fixed failure redirection

+++ Assets/Scripts/CoreManager.cs
@@ -18,7 +18,7 @@ public class CoreManager : MonoBehaviour
     private List<Character> knownChars = new List<Character>();
     [NonSerialized] public Character currentCharacter;
     [NonSerialized] public float minigameScore;
-    [NonSerialized] public int dialogueScore;
+    [NonSerialized] public float dialogueScore;
     [NonSerialized] public bool hasPlayedMinigame;
     
     // This should only be loaded in once, after the start screen. When it is, we want to go straight to the
